Improve UserDAO logic

+++ src/daos/UserDAO.java
@@ -12,20 +12,51 @@ import java.util.UUID;
 public final class UserDAO {
     private static final String fileName = "users.csv";
     private static final String[] headers = {"UUID", "FullName", "Email", "Password"};
-    public static final CSVFormat CSV_FORMAT = CSVFormat.DEFAULT.builder()
-            .setHeader(headers)
-            .get();
+    private static CSVFormat CSV_FORMAT;
 
     private UserDAO() {}
 
+    private static boolean fileExists() {
+        File file = new File(fileName);
+        try {
+            if (!file.exists()) {
+                file.createNewFile();
+                CSV_FORMAT = CSVFormat.DEFAULT.builder().setHeader(headers).get();
+            } else {
+                CSV_FORMAT = CSVFormat.DEFAULT.builder().get();
+            }
+        } catch (IOException e) {
+            Display.error("Something went wrong: " + e.getMessage());
+            return false;
+        }
+        return true;
+    }
+
     public static void saveUser(String fullName, String email, String password) {
+        if (!fileExists()) {
+            return;
+        }
+
         try (
                 Writer writer = new FileWriter(fileName, true);
-                CSVPrinter csvPrinter = new CSVPrinter(writer, CSV_FORMAT)
+                CSVPrinter csvPrinter = new CSVPrinter(writer, CSV_FORMAT);
+                Reader reader = new FileReader(fileName)
         ){
+            Iterable<CSVRecord> records = CSVFormat.DEFAULT.builder()
+                    .setSkipHeaderRecord(true)
+                    .setHeader(headers)
+                    .build()
+                    .parse(reader);
+
+            for (CSVRecord record : records) {
+                if (email.equals(record.get("Email"))) {
+                    Display.error("User with this email already exists.");
+                    return;
+                }
+            }
+
             String uuid = UUID.randomUUID().toString();
             csvPrinter.printRecord(uuid, fullName, email, password);
-            Display.success("User Saved, you can log in now!");
             AccountDAO.createAccount(uuid);
         } catch (IOException e) {
             Display.error("Something went wrong: " + e.getMessage());
@@ -33,6 +64,10 @@ public final class UserDAO {
     }
 
     public static void loginUser(String email, String password) {
+        if (!fileExists()) {
+            return;
+        }
+
         try (Reader reader = new FileReader(fileName)) {
             Iterable<CSVRecord> records = CSVFormat.DEFAULT.builder()
                     .setSkipHeaderRecord(true)
